% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tephra3D.R
\name{BlastSim3d}
\alias{BlastSim3d}
\title{Explosive Tephra Dispersion Model}
\usage{
BlastSim3d(v, th_i = 2 * 1:40, th_a = 0, dt = 0.01, ...)
}
\arguments{
\item{v}{initial velocity [m/s]}

\item{th_i}{initial inclination angles to test [degrees]}

\item{th_a}{initial azimuth angles to test [degrees clockwise from north]}

\item{dt}{model time interval [s]}

\item{...}{parameters to pass to Tephra3d}
}
\description{
Models the transport of particles ejected at the same velocity and different angles using tephra3d.
}
\examples{
BlastSim3d(v = 10, th_i = 2* 1:40, th_a = 0, dt = 0.01)
}
\keyword{misc}
