% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshgridn.R
\name{meshgridn}
\alias{meshgridn}
\title{N-Dimensional Meshgrid
Make a list in which all combinations of inputs occur.}
\usage{
meshgridn(L)
}
\arguments{
\item{L}{list containing any number of vectors}
}
\description{
N-Dimensional Meshgrid
Make a list in which all combinations of inputs occur.
}
\examples{
meshgridn(list(1:2, 5:6, 7:9))
}
\keyword{misc}
